+++ config.html
@@ -1,0 +1,19 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Vite App</title>
+  <link rel="stylesheet" href="/src/main.css" />
+  <script type="module" src="/src/config.js"></script>
+</head>
+
+<body>
+    <h1>Config</h1>
+      <ul id="datesList"></ul>
+      <button id="resetButton">Reset Drawn Dates</button>
+</body>
+
+</html>
