--- dates.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Drawn Dates</title>
-  <style>
-    body {
-      font-family: Arial, sans-serif;
-      margin: 10px;
-    }
-    ul {
-      list-style-type: none;
-      padding: 0;
-    }
-    li {
-      margin: 5px 0;
-      font-size: 14px;
-    }
-  </style>
-</head>
-<body>
-  <h1>Drawn Dates</h1>
-  <ul id="datesList"></ul>
-  <button id="resetButton">Reset Drawn Dates</button>
-  <script src="dates.js"></script>
-</body>
-</html>
-
