+++ popup.html
@@ -1,57 +1,39 @@
-<!DOCTYPE html>
-<html>
+<!doctype html>
+<html lang="en">
+
 <head>
-  <title>Random Game Launcher</title>
-  <style>
-    body {
-      font-family: Arial, sans-serif;
-      width: 300px;
-      margin: 10px;
-    }
-    label, button {
-      font-size: 14px;
-      display: block;
-      margin: 10px 0;
-    }
-    input {
-      width: 100%;
-      padding: 5px;
-      margin-bottom: 10px;
-      font-size: 14px;
-    }
-    button {
-      padding: 10px;
-      font-size: 14px;
-      background-color: #4CAF50;
-      color: white;
-      border: none;
-      cursor: pointer;
-    }
-    button:hover {
-      background-color: #45a049;
-    }
-    .horizontal {
-      display: flex;
-    }
-    .horizontal button {
-      margin: 0 5px;
-    }
-  </style>
+  <meta charset="UTF-8" />
+  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Vite App</title>
+  <link rel="stylesheet" href="/src/main.css" />
 </head>
-<body>
-  <h1>Random Game</h1>
-  <label for="startDate">Start Date:</label>
-  <input type="date" id="startDate" value="2024-01-01" />
 
-  <label for="endDate">End Date:</label>
-  <input type="date" id="endDate" value="2024-12-31" />
+<body class="p-4">
+  <script type="module" src="/src/popup.js"></script>
 
-  <div class="horizontal">
-    <button id="playButton">Play a Random Game</button>
-    <button id="viewDatesButton">View Drawn Dates</button>
-  </div>
+  <h1 class="text-3xl mb-4 font-extrabold leading-none text-gray-900 tracking-tight">Random Game</h1>
 
-  <script src="popup.js"></script>
+  <form class="w-64">
+    <div class="flex gap-2 justify-between">
+      <div>
+        <label class="block text-gray-700" for="startDate">Start Date:</label>
+        <input class="border rounded text-sm p-1" type="date" id="startDate" value="2024-01-01" />
+      </div>
+
+      <div>
+        <label class="block text-gray-700" for="endDate">End Date:</label>
+        <input class="border rounded text-sm p-1" type="date" id="endDate" value="2024-12-31" />
+      </div>
+    </div>
+
+    <button class="w-full rounded bg-green-600 p-3 my-2 font-bold text-lg text-white" id="playButton">Play a Random Game</button>
+    <span class="text-xs text-gray-300">
+      Developed by <a class="underline transition-colors hover:text-transparent bg-clip-text bg-gradient-to-r to-emerald-600 from-sky-400" href="https://github.com/Northo/ny-launcher">Northo</a>.</span>
+    <button class="text-gray-600 text-xs ml-3 mt-1 float-right" id="viewDatesButton">Setup</button>
+  </form>
+
+
 </body>
-</html>
 
+</html>
