+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
   "manifest_version": 3,
-  "name": "Play Mini!",
+  "name": "Play Mini (vite version)!",
   "version": "1.2.0",
   "description": "Launch a random NY mini.",
   "permissions": ["tabs", "storage"],
