+++ tailwind.config.js
@@ -1,0 +1,9 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ["./*.html", "./src/**/*.{js,html}"],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
+
