+++ vite.config.js
@@ -1,0 +1,13 @@
+import { defineConfig } from "vite"
+
+export default defineConfig({
+    build: {
+      rollupOptions: {
+        // https://rollupjs.org/configuration-options/
+        input: {
+            popup: 'popup.html',
+            config: 'config.html'
+        }
+      },
+    },
+  })
